data-archive/requirements.txt
psycopg2-binary==2.9.6
configparser==5.2.0
datetime==5.2.0
